move splash activity and add themeindex collect, fix weeklyviewmodel to weeklysharedviewmodel by activityviewmodels, etc
