+++ app/src/main/java/com/moonwinston/motivationaltodolist/ui/daily/DailyViewModel.kt
@@ -1,13 +1,13 @@
 package com.moonwinston.motivationaltodolist.ui.daily
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.moonwinston.motivationaltodolist.data.AchievementRateRepository
 import com.moonwinston.motivationaltodolist.data.TaskEntity
+import com.moonwinston.motivationaltodolist.data.TaskRepository
 import com.moonwinston.motivationaltodolist.data.UserPreferencesRepository
 import com.moonwinston.motivationaltodolist.utils.dateOfToday
+import com.moonwinston.motivationaltodolist.utils.getDateExceptTime
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
@@ -15,19 +15,22 @@ import javax.inject.Inject
 
 @HiltViewModel
 class DailyViewModel @Inject constructor(
+    private val taskRepository: TaskRepository,
     private val achievementRateRepository: AchievementRateRepository,
     private val userPreferencesRepository: UserPreferencesRepository
 ) : ViewModel() {
 
-    fun calculateRate(tasksList: List<TaskEntity>): Float {
-        var totalTasks = 0F
-        var doneTasks = 0F
-        for (task in tasksList) {
-            totalTasks += 1F
-            if (task.isCompleted) doneTasks += 1F
+    val todayTaskLists = taskRepository.getAllTasks().map { taskEntities ->
+        taskEntities.filter { taskEntity ->
+            taskEntity.taskDate.getDateExceptTime() == dateOfToday()
+        }.sortedBy { taskEntity ->
+            taskEntity.taskDate
         }
-        return if (doneTasks == 0F) 0F else doneTasks / totalTasks
-    }
+    }.stateIn(
+        initialValue = emptyList(),
+        started = SharingStarted.Eagerly,
+        scope = viewModelScope
+    )
 
     val todayAchievementRate =
         achievementRateRepository.getRateByDates(dateOfToday()).filterNotNull().stateIn(
@@ -47,5 +50,15 @@ class DailyViewModel @Inject constructor(
         viewModelScope.launch {
             userPreferencesRepository.updateDailyCoachMarkDismissedStatusFlow(dismissDailyCoachMark)
         }
+    }
+
+    fun calculateRate(tasksList: List<TaskEntity>): Float {
+        var totalTasks = 0F
+        var doneTasks = 0F
+        for (task in tasksList) {
+            totalTasks += 1F
+            if (task.isCompleted) doneTasks += 1F
+        }
+        return if (doneTasks == 0F) 0F else doneTasks / totalTasks
     }
 }
