+++ app/src/main/java/com/moonwinston/motivationaltodolist/ui/splash/SplashActivity.kt
@@ -1,4 +1,4 @@
-package com.moonwinston.motivationaltodolist
+package com.moonwinston.motivationaltodolist.ui.splash
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,6 +8,7 @@ import androidx.lifecycle.lifecycleScope
 import com.moonwinston.motivationaltodolist.databinding.ActivitySplashBinding
 import com.moonwinston.motivationaltodolist.ui.main.MainActivity
 import com.moonwinston.motivationaltodolist.ui.main.MainViewModel
+import com.moonwinston.motivationaltodolist.utils.setNightMode
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.*
 
@@ -18,9 +19,13 @@ class SplashActivity : AppCompatActivity() {
     
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-//        ThemeUtil().setTheme(mainViewModel.themeIndex.value)
         binding = ActivitySplashBinding.inflate(layoutInflater)
         setContentView(binding.root)
+        lifecycleScope.launch {
+            mainViewModel.themeIndex.collect { themeIndex ->
+                setNightMode(themeIndex)
+            }
+        }
         lifecycleScope.launch(Dispatchers.Main) {
             delay(2000)
             val intent = Intent(this@SplashActivity, MainActivity::class.java).apply {
